+++ src/utils/time_mod.F90
@@ -38,17 +38,17 @@ module time_mod
   end type alert_type
 
   ! Namelist parameters
-  integer, public :: start_time_array(5)  = 0
-  integer, public :: end_time_array(5)    = 0
-  real(8), public :: run_hours            = 0
-  real(8), public :: run_days             = 0
-  real(8), public :: run_years            = 0
+  integer , public :: start_time_array(5)  = 0
+  integer , public :: end_time_array(5)    = 0
+  real(r8), public :: run_hours            = 0
+  real(r8), public :: run_days             = 0
+  real(r8), public :: run_years            = 0
 
   type(datetime_type) start_time
   type(datetime_type) end_time
   type(datetime_type) curr_time
   type(timedelta_type) dt
-  real(8) elapsed_seconds
+  real(r8) elapsed_seconds
   type(hash_table_type) alerts
   integer time_step
   integer old_time_idx
@@ -60,7 +60,7 @@ contains
 
   subroutine time_init(dt_in_seconds)
 
-    real(8), intent(in) :: dt_in_seconds
+    real(r8), intent(in) :: dt_in_seconds
 
     if (sum(start_time_array) > 0) then
       call start_time%init(year=start_time_array(1),  &
@@ -127,7 +127,7 @@ contains
 
   subroutine time_advance(dt_in_seconds)
 
-    real(8), intent(in), optional :: dt_in_seconds
+    real(r8), intent(in), optional :: dt_in_seconds
 
     type(hash_table_iterator_type) iter
 
@@ -160,7 +160,7 @@ contains
 
   subroutine time_fast_forward(time_value, time_units)
 
-    real(8), intent(in) :: time_value
+    real(r8), intent(in) :: time_value
     character(*), intent(in) :: time_units
 
     type(timedelta_type) skipped_time
@@ -203,7 +203,7 @@ contains
 
   end subroutine time_fast_forward
 
-  real(8) function time_elapsed_seconds() result(res)
+  real(r8) function time_elapsed_seconds() result(res)
 
     res = elapsed_seconds
 
@@ -224,17 +224,17 @@ contains
   subroutine time_add_alert(name, months, days, hours, minutes, seconds)
 
     character(*), intent(in)           :: name
-    real(8)     , intent(in), optional :: months
-    real(8)     , intent(in), optional :: days
-    real(8)     , intent(in), optional :: hours
-    real(8)     , intent(in), optional :: minutes
-    real(8)     , intent(in), optional :: seconds
+    real(r8)    , intent(in), optional :: months
+    real(r8)    , intent(in), optional :: days
+    real(r8)    , intent(in), optional :: hours
+    real(r8)    , intent(in), optional :: minutes
+    real(r8)    , intent(in), optional :: seconds
 
-    real(8) months_
-    real(8) days_
-    real(8) hours_
-    real(8) minutes_
-    real(8) seconds_
+    real(r8) months_
+    real(r8) days_
+    real(r8) hours_
+    real(r8) minutes_
+    real(r8) seconds_
     type(alert_type) alert
 
     if (present(months)) then
